Separate component for governments page, bug fixes
