+++ src_ts/components/pages/partners/pages/list/partners-filters.ts
@@ -1,5 +1,5 @@
 import {AnyObject} from '@unicef-polymer/etools-types/dist/global.types';
-import {translate} from 'lit-translate';
+import {translate, get as getTranslation} from 'lit-translate';
 import {setselectedValueTypeByFilterKey} from '@unicef-polymer/etools-modules-common/dist/list/filters';
 import {EtoolsFilterTypes} from '@unicef-polymer/etools-modules-common/dist/layout/filters/etools-filters';
 
@@ -27,87 +27,80 @@ export const selectedValueTypeByFilterKey: AnyObject = {
 
 setselectedValueTypeByFilterKey(selectedValueTypeByFilterKey);
 
-let showOnlyGovernmentType = false;
-export function setShowOnlyGovernmentType(val: boolean) {
-  showOnlyGovernmentType = val;
-}
-
-export function getPartnerFilters() {
-  return [
-    {
-      filterName: translate('GENERAL.SEARCH_RECORDS'),
-      filterKey: PartnerFilterKeys.search,
-      type: EtoolsFilterTypes.Search,
-      selectedValue: '',
-      selected: true
-    },
-    {
-      filterName: translate('PARTNER_TYPE'),
-      filterKey: PartnerFilterKeys.partner_types,
-      type: EtoolsFilterTypes.DropdownMulti,
-      selectionOptions: [],
-      selectedValue: [],
-      selected: true,
-      minWidth: '350px',
-      hideSearch: true,
-      disabled: showOnlyGovernmentType
-    },
-    {
-      filterName: translate('CSO_TYPE'),
-      filterKey: PartnerFilterKeys.cso_types,
-      type: EtoolsFilterTypes.DropdownMulti,
-      selectionOptions: [],
-      selectedValue: [],
-      selected: true,
-      minWidth: '350px',
-      hideSearch: false,
-      disabled: showOnlyGovernmentType
-    },
-    {
-      filterName: translate('HACT_RISK_RATING'),
-      filterKey: PartnerFilterKeys.risk_ratings,
-      type: EtoolsFilterTypes.DropdownMulti,
-      selectionOptions: [],
-      selectedValue: [],
-      selected: true,
-      minWidth: '160px',
-      hideSearch: false,
-      disabled: false
-    },
-    {
-      filterName: translate('SEA_RISK_RATING'),
-      filterKey: PartnerFilterKeys.sea_risk_ratings,
-      type: EtoolsFilterTypes.DropdownMulti,
-      selectionOptions: [],
-      selectedValue: [],
-      selected: false,
-      minWidth: '160px',
-      hideSearch: true,
-      disabled: false
-    },
-    {
-      filterName: translate('PSEA_ASSESSMENT_DATE_BEFORE'),
-      filterKey: PartnerFilterKeys.psea_assessment_date_before,
-      type: EtoolsFilterTypes.Date,
-      selectedValue: '',
-      path: 'selectedPseaDateBefore',
-      selected: false,
-      disabled: false
-    },
-    {
-      filterName: translate('PSEA_ASSESSMENT_DATE_AFTER'),
-      filterKey: PartnerFilterKeys.psea_assessment_date_after,
-      type: EtoolsFilterTypes.Date,
-      selectedValue: '',
-      selected: false,
-      disabled: false
-    },
-    {
-      filterName: translate('SHOW_HIDDEN'),
-      filterKey: PartnerFilterKeys.hidden,
-      type: EtoolsFilterTypes.Toggle,
-      selectedValue: false,
-      selected: true
-    }
-  ];
-}
+export const partnerFilters = [
+  {
+    filterName: getTranslation('GENERAL.SEARCH_RECORDS'),
+    filterKey: PartnerFilterKeys.search,
+    type: EtoolsFilterTypes.Search,
+    selectedValue: '',
+    selected: true
+  },
+  {
+    filterName: getTranslation('PARTNER_TYPE'),
+    filterKey: PartnerFilterKeys.partner_types,
+    type: EtoolsFilterTypes.DropdownMulti,
+    selectionOptions: [],
+    selectedValue: [],
+    selected: true,
+    minWidth: '350px',
+    hideSearch: true,
+    disabled: false
+  },
+  {
+    filterName: getTranslation('CSO_TYPE'),
+    filterKey: PartnerFilterKeys.cso_types,
+    type: EtoolsFilterTypes.DropdownMulti,
+    selectionOptions: [],
+    selectedValue: [],
+    selected: true,
+    minWidth: '350px',
+    hideSearch: false,
+    disabled: false
+  },
+  {
+    filterName: getTranslation('HACT_RISK_RATING'),
+    filterKey: PartnerFilterKeys.risk_ratings,
+    type: EtoolsFilterTypes.DropdownMulti,
+    selectionOptions: [],
+    selectedValue: [],
+    selected: true,
+    minWidth: '160px',
+    hideSearch: false,
+    disabled: false
+  },
+  {
+    filterName: getTranslation('SEA_RISK_RATING'),
+    filterKey: PartnerFilterKeys.sea_risk_ratings,
+    type: EtoolsFilterTypes.DropdownMulti,
+    selectionOptions: [],
+    selectedValue: [],
+    selected: false,
+    minWidth: '160px',
+    hideSearch: true,
+    disabled: false
+  },
+  {
+    filterName: getTranslation('PSEA_ASSESSMENT_DATE_BEFORE'),
+    filterKey: PartnerFilterKeys.psea_assessment_date_before,
+    type: EtoolsFilterTypes.Date,
+    selectedValue: '',
+    path: 'selectedPseaDateBefore',
+    selected: false,
+    disabled: false
+  },
+  {
+    filterName: getTranslation('PSEA_ASSESSMENT_DATE_AFTER'),
+    filterKey: PartnerFilterKeys.psea_assessment_date_after,
+    type: EtoolsFilterTypes.Date,
+    selectedValue: '',
+    selected: false,
+    disabled: false
+  },
+  {
+    filterName: getTranslation('SHOW_HIDDEN'),
+    filterKey: PartnerFilterKeys.hidden,
+    type: EtoolsFilterTypes.Toggle,
+    selectedValue: false,
+    selected: true
+  }
+];
